requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
awsebcli==3.19.2
bcrypt==3.2.0
blessed==1.17.11
boto3==1.16.23
botocore==1.19.23
cached-property==1.5.2
cement==2.8.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.0
Django==3.1.3
django-environ==0.4.5
django-model-utils==4.0.0
django-private-files==1.0.1
django-storages==1.10.1
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jmespath==0.10.0
jsonschema==3.2.0
msgpack==1.0.0
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.0.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
semantic-version==2.5.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.57.0
zope.interface==5.1.2
